Packing functions
 - Added a random position generator, and functions to place several persons on a terrain.
 - Added a sample terrains with 2 big obstacles for testing purposes

TODO : no chek on the number of available slots prior to placing terrains :  the program will go into an infinite loop if there are not enough slots for the number of persons to place.
TODO : no check on obstacles adding : on a small terrain, some out of range errors can happen.